requirements.txt
Pillow
matplotlib

